+++ src/__tests__/client/dynamic.test.ts
@@ -428,7 +428,7 @@ describe('DynamicFeature', () => {
         }
       }, ['*'])
       await createFeature.send({ files: [URI.file(os.tmpdir()), URI.file(__filename)] })
-      await helper.waitValue(() => n, 1)
+      await helper.waitValue(() => n, 2)
       await client.stop()
     })
   })